requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
google-generativeai==0.3.2
pillow==10.2.0
pdf2image==1.17.0
python-dotenv==1.0.1
sqlalchemy==2.0.23
pymysql==1.1.0
pydantic-settings==2.1.0

